refact: Extract ActivateScopes statement for reuse
