+++ core/src/main/java/io/github/cdiunit/CdiRunner.java
@@ -15,6 +15,8 @@
  */
 package io.github.cdiunit;
 
+import java.util.concurrent.atomic.AtomicBoolean;
+
 import javax.naming.InitialContext;
 
 import org.jboss.weld.environment.se.Weld;
@@ -28,7 +30,7 @@ import org.junit.runners.model.Statement;
 
 import io.github.cdiunit.internal.TestConfiguration;
 import io.github.cdiunit.internal.WeldHelper;
-import io.github.cdiunit.internal.activatescopes.ScopesHelper;
+import io.github.cdiunit.internal.junit4.ActivateScopes;
 
 /**
  * <code>&#064;CdiRunner</code> is a JUnit runner that uses a CDI container to
@@ -50,14 +52,15 @@ import io.github.cdiunit.internal.activatescopes.ScopesHelper;
  */
 public class CdiRunner extends BlockJUnit4ClassRunner {
 
+    private static final String JNDI_FACTORY_PROPERTY = "java.naming.factory.initial";
+
     private Class<?> clazz;
     private Weld weld;
     private WeldContainer container;
     private Throwable startupException;
     private FrameworkMethod frameworkMethod;
     private TestConfiguration testConfiguration;
-    private static final String JNDI_FACTORY_PROPERTY = "java.naming.factory.initial";
-    private boolean contextsActivated;
+    private final AtomicBoolean contextsActivated = new AtomicBoolean();
 
     public CdiRunner(Class<?> clazz) throws InitializationError {
         super(clazz);
@@ -128,7 +131,7 @@ public class CdiRunner extends BlockJUnit4ClassRunner {
     protected Statement methodBlock(final FrameworkMethod frameworkMethod) {
         this.frameworkMethod = frameworkMethod;
         var statement = super.methodBlock(frameworkMethod);
-        statement = new CdiContextStatement(statement);
+        statement = new ActivateScopes(statement, testConfiguration, contextsActivated, () -> container.getBeanManager());
         final var defaultStatement = statement;
         return new Statement() {
 
@@ -166,34 +169,6 @@ public class CdiRunner extends BlockJUnit4ClassRunner {
 
             }
         };
-
-    }
-
-    class CdiContextStatement extends Statement {
-
-        private final Statement next;
-
-        CdiContextStatement(Statement next) {
-            this.next = next;
-        }
-
-        @Override
-        public void evaluate() throws Throwable {
-            final var method = testConfiguration.getTestMethod();
-            final var isolationLevel = testConfiguration.getIsolationLevel();
-            try {
-                if (!contextsActivated) {
-                    ScopesHelper.activateContexts(container.getBeanManager(), method);
-                    contextsActivated = true;
-                }
-                next.evaluate();
-            } finally {
-                if (contextsActivated && isolationLevel == IsolationLevel.PER_METHOD) {
-                    contextsActivated = false;
-                    ScopesHelper.deactivateContexts(container.getBeanManager(), method);
-                }
-            }
-        }
 
     }
