+++ core/src/main/java/io/github/cdiunit/internal/junit4/package-info.java
@@ -1,0 +1,1 @@
+package io.github.cdiunit.internal.junit4;
